% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_sum.R
\name{adapt_sum}
\alias{adapt_sum}
\title{Test function}
\usage{
adapt_sum(
  lFun,
  params = numeric(),
  eps = 1e-15,
  maxIter = 1e+05,
  logL = -Inf,
  n0 = 0
)
}
\description{
Wraps the adapt_sum C function for testing
}
