% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerics.R
\name{log_diff_exp}
\alias{log_diff_exp}
\title{Log of exp(x)-exp(y)}
\usage{
log_diff_exp(x, y)
}
\arguments{
\item{x}{a scalar}

\item{y}{a scalar}
}
\value{
log(exp(x)-exp(y))
}
\description{
Log of exp(x)-exp(y)
}
