% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerics.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Log of sum of exponentials}
\usage{
log_sum_exp(x)
}
\arguments{
\item{x}{a vector with logarithms of the values whose sum will be computed.}
}
\value{
the log of the sum of the exponentials of the elements in x.
}
\description{
Log of sum of exponentials
}
