% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerics.R
\name{relative_difference}
\alias{relative_difference}
\title{Relative difference}
\usage{
relative_difference(x, y)
}
\arguments{
\item{x}{a scalar.}

\item{y}{a scalar.}
}
\value{
relative difference, as measured by \code{all.equal} with \code{tolerance = 0}.
}
\description{
Relative difference
}
