% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerics.R
\name{robust_difference}
\alias{robust_difference}
\title{Returns the difference between exp(x) and exp(y)}
\usage{
robust_difference(x, y, log = FALSE)
}
\arguments{
\item{x}{a scalar.}

\item{y}{a scalar.}

\item{log}{logical. If \code{TRUE}, the logarithmic of the difference is returned. Default is \code{FALSE}.}
}
\value{
log( max(exp(x), exp(y)) - min(exp(x), exp(y)) )
}
\description{
Returns the difference between exp(x) and exp(y)
}
